refactor: split groups out from users in security
