+++ lib/services/security/service.js
@@ -270,7 +270,7 @@ SecurityService.prototype.__ensureAdminUser = function (config, callback) {
       if (!config.adminUser.password) config.adminUser.password = 'happn';
 
       //group, options, callback
-      _this.users.__upsertGroup(config.adminGroup, {}, function (e, adminGroup) {
+      _this.groups.__upsertGroup(config.adminGroup, {}, function (e, adminGroup) {
 
         if (e) return reject(e);
 
@@ -278,7 +278,7 @@ SecurityService.prototype.__ensureAdminUser = function (config, callback) {
 
           if (e) return reject(e);
 
-          _this.users.linkGroup(adminGroup, adminUser, function(e){
+          _this.groups.linkGroup(adminGroup, adminUser, function(e){
             if (e) return reject(e);
             resolve();
           });
@@ -296,6 +296,8 @@ SecurityService.prototype.initialize = function (config, callback) {
     this.dataService = this.happn.services.data;
     this.cryptoService = this.happn.services.crypto;
     this.sessionService =  this.happn.services.session;
+    this.utilsService =  this.happn.services.utils;
+    this.errorService =  this.happn.services.error;
 
     this.pathField = this.dataService.pathField; //backward compatible for allowing mongo plugin, which uses an actual path field
 
@@ -325,7 +327,10 @@ SecurityService.prototype.initialize = function (config, callback) {
 
     var _this = this;
 
-    _this.__initializeUsers(config)
+    _this.__initializeGroups(config)
+      .then(function(){
+        return _this.__initializeUsers(config);
+      })
       .then(function(){
         return _this.__initializeProfiles(config);
       })
@@ -387,7 +392,34 @@ SecurityService.prototype.__initializeUsers = function (config) {
       value: _this.happn
     });
 
+    Object.defineProperty(_this.users, 'groups', {
+      value: _this.groups
+    });
+
     _this.users.initialize(config, _this, function(e){
+      if (e) return reject(e);
+      resolve();
+    });
+  });
+};
+
+SecurityService.prototype.__initializeGroups = function (config) {
+
+  var _this = this;
+
+  return new Promise(function(resolve, reject){
+
+    var SecurityGroups = require('./groups');
+
+    _this.groups = new SecurityGroups({
+      logger: _this.log
+    });
+
+    Object.defineProperty(_this.groups, 'happn', {
+      value: _this.happn
+    });
+
+    _this.groups.initialize(config, _this, function(e){
       if (e) return reject(e);
       resolve();
     });
@@ -843,6 +875,9 @@ SecurityService.prototype.dataChanged = function (whatHappnd, changedData, addit
   var _this = this;
 
   _this.users.clearCaches()
+    .then(function(){
+      return _this.groups.clearCaches();
+    })
     .then(function(){
       return _this.resetSessionPermissions(whatHappnd, changedData, additionalInfo);
     })
@@ -1316,6 +1351,57 @@ SecurityService.prototype.stop = function (options, callback) {
   if (this.__cache_revoked_sessions) this.__cache_revoked_sessions.clear();
 
   callback();
+};
+
+SecurityService.prototype.validateName = function (name, validationType) {
+
+  if (!name) throw new this.errorService.ValidationError('names cannot be empty');
+
+  if (this.utilsService.stringContainsAny(name, ['/', '_SYSTEM', '_GROUP', '_PERMISSION', '_USER_GROUP', '_ADMIN']))
+    throw new this.errorService.ValidationError('validation error: ' + validationType + ' names cannot contain the special _SYSTEM, _GROUP, _PERMISSION, _USER_GROUP, _ADMIN segment or a forward slash /');
+
+};
+
+SecurityService.prototype.checkOverwrite = function (validationType, obj, path, name, options, callback) {
+
+  if (!name) name = obj.name;
+
+  if (options && options.overwrite === false) {
+
+    this.dataService.get(path, {}, function (e, result) {
+
+      if (e) return callback(e);
+
+      if (result) return callback(new Error('validation failure: ' + validationType + ' by the name ' + name + ' already exists'));
+
+      callback();
+    });
+  } else return callback();
+};
+
+SecurityService.prototype.serializeAll = function (objectType, objArray, options) {
+
+  var _this = this;
+
+  if (!objArray) return [];
+
+  return objArray
+
+    .map(function (obj) {
+      return _this.serialize(objectType, obj, options);
+    })
+
+};
+
+SecurityService.prototype.serialize = function (objectType, obj) {
+
+  var returnObj = this.utilsService.clone(obj.data);
+
+  returnObj._meta = obj._meta;
+
+  if (objectType == 'user') delete returnObj['password'];
+
+  return returnObj;
 };
 
 SecurityService.prototype.safe = function () {
