+++ lib/services/security/users.js
@@ -24,13 +24,6 @@ SecurityUsers.prototype.initialize = Promise.promisify(function (config, securit
     _this.cryptoService = _this.happn.services.crypto;
     _this.sessionService = _this.happn.services.session;
 
-    if (!config.__cache_groups) config.__cache_groups = {
-      type: 'LRU',
-      cache: {
-        max: 1000,
-        maxAge: 0
-      }
-    };
     if (!config.__cache_users) config.__cache_users = {
       type: 'LRU',
       cache: {
@@ -39,9 +32,18 @@ SecurityUsers.prototype.initialize = Promise.promisify(function (config, securit
       }
     };
 
-    _this.__cache_groups = _this.cacheService.new('cache_security_groups', config.__cache_groups);
     _this.__cache_users = _this.cacheService.new('cache_security_users', config.__cache_users);
     _this.__cache_passwords = _this.cacheService.new('cache_security_passwords', config.__cache_users);
+
+
+    //backward compatibility after taking groups methods out
+
+    _this.unlinkGroup = _this.groups.unlinkGroup.bind(_this.groups);
+    _this.linkGroup = _this.groups.linkGroup.bind(_this.groups);
+    _this.getGroup = _this.groups.getGroup.bind(_this.groups);
+    _this.listGroups = _this.groups.listGroups.bind(_this.groups);
+    _this.deleteGroup = _this.groups.deleteGroup.bind(_this.groups);
+    _this.upsertGroup = _this.groups.upsertGroup.bind(_this.groups);
 
     callback();
 
@@ -57,69 +59,14 @@ SecurityUsers.prototype.clearCaches = function () {
   return _this.__cache_passwords.clear()
     .then(function(){
       return _this.__cache_users.clear();
-    })
-    .then(function(){
-      return _this.__cache_groups.clear();
-    })
-};
-
-SecurityUsers.prototype.__validateName = function (name, validationType) {
-
-  if (!name) throw new this.errorService.ValidationError('names cannot be empty');
-
-  if (this.utilsService.stringContainsAny(name, ['/', '_SYSTEM', '_GROUP', '_PERMISSION', '_USER_GROUP', '_ADMIN']))
-    throw new this.errorService.ValidationError('validation error: ' + validationType + ' names cannot contain the special _SYSTEM, _GROUP, _PERMISSION, _USER_GROUP, _ADMIN segment or a forward slash /');
-
-};
-
-SecurityUsers.prototype.__checkOverwrite = function (validationType, obj, path, name, options, callback) {
-
-  if (!name) name = obj.name;
-
-  if (options && options.overwrite === false) {
-
-    this.dataService.get(path, {}, function (e, result) {
-
-      if (e) return callback(e);
-
-      if (result) return callback(new Error('validation failure: ' + validationType + ' by the name ' + name + ' already exists'));
-
-      callback();
     });
-  } else return callback();
 };
 
 SecurityUsers.prototype.__validate = function (validationType, options, obj, callback) {
 
   var _this = this;
 
-  if (validationType == 'user-group') {
-
-    var group = options[0];
-    var user = options[1];
-
-    if (!group._meta) return callback(new Error('validation error: group does not exist or has not been saved'));
-
-    if (!user._meta) return callback(new Error('validation error: user does not exist or has not been saved'));
-
-    return _this.dataService.get(group._meta.path, {}, function (e, result) {
-
-      if (e) return callback(e);
-
-      if (!result) return callback(new Error('validation error: group does not exist: ' + group._meta.path));
-
-      return _this.dataService.get(user._meta.path, {}, function (e, result) {
-
-        if (e) return callback(e);
-
-        if (!result) return callback(new Error('validation error: user does not exist: ' + user._meta.path));
-
-        callback();
-      });
-    });
-  }
-
-  if (obj.name) _this.__validateName(obj.name, validationType);
+  if (obj.name) this.securityService.validateName(obj.name, validationType);
 
   if (validationType == 'user') {
 
@@ -133,47 +80,10 @@ SecurityUsers.prototype.__validate = function (validationType, options, obj, cal
 
         if (!result) return callback(new Error('validation failure: no password or publicKey specified for a new user'));
 
-        _this.__checkOverwrite(validationType, obj, '/_SYSTEM/_SECURITY/_USER/' + obj.username, obj.username, options, callback);
+        _this.securityService.checkOverwrite(validationType, obj, '/_SYSTEM/_SECURITY/_USER/' + obj.username, obj.username, options, callback);
       });
     }
-    return _this.__checkOverwrite(validationType, obj, '/_SYSTEM/_SECURITY/_USER/' + obj.username, obj.username, options, callback);
-  }
-
-  if (validationType == 'group') {
-
-    if (options.parent) {
-
-      if (!options.parent._meta.path) return callback(new Error('validation error: parent group path is not in your request, have you included the _meta?'));
-      //path, parameters, callback
-      return _this.dataService.get(options.parent._meta.path, {}, function (e, result) {
-
-        if (e) return callback(e);
-
-        if (!result) return callback(new Error('validation error: parent group does not exist: ' + options.parent._meta.path));
-
-        _this.__checkOverwrite(validationType, obj, options.parent._meta.path + '/' + obj.name, obj.name, options, callback);
-
-      });
-    }
-    return _this.__checkOverwrite(validationType, obj, '/_SYSTEM/_SECURITY/_GROUP/' + obj.name, obj.name, options, callback);
-  }
-
-  if (validationType == 'permission') {
-
-    var permission = options[0];
-    var permissionGroup = options[1];
-
-    if (!permissionGroup) return callback(new Error('validation error: you need a group to add a permission to'));
-    if (!permissionGroup._meta.path) return callback(new Error('validation error: permission group path is not in your request, have you included the _meta?'));
-
-    return _this.dataService.get(permissionGroup._meta.path, {}, function (e, result) {
-
-      if (e) return callback(e);
-
-      if (!result) return callback(new Error('validation error: permission group does not exist: ' + permissionGroup._meta.path));
-
-      callback();
-    });
+    return _this.securityService.checkOverwrite(validationType, obj, '/_SYSTEM/_SECURITY/_USER/' + obj.username, obj.username, options, callback);
   }
 
   return callback(new Error('Unknown validation type: ' + validationType));
@@ -202,33 +112,8 @@ SecurityUsers.prototype.getPasswordHash = function (username, callback) {
         callback(null, user.data.password);
       });
 
-    }.bind(this));
+    });
   });
-};
-
-SecurityUsers.prototype.serializeAll = function (objectType, objArray, options) {
-
-  var _this = this;
-
-  if (!objArray) return [];
-
-  return objArray
-
-    .map(function (obj) {
-      return _this.serialize(objectType, obj, options);
-    })
-
-};
-
-SecurityUsers.prototype.serialize = function (objectType, obj) {
-
-  var returnObj = this.utilsService.clone(obj.data);
-
-  returnObj._meta = obj._meta;
-
-  if (objectType == 'user') delete returnObj['password'];
-
-  return returnObj;
 };
 
 /**
@@ -274,7 +159,7 @@ SecurityUsers.prototype.__upsertUser = function (user, options, callback) {
 
       if (e) return callback(e);
 
-      var upserted = _this.serialize('user', result);
+      var upserted = _this.securityService.serialize('user', result);
 
       _this.securityService.dataChanged('upsert-user', upserted, prepared, function(){
 
@@ -294,8 +179,8 @@ SecurityUsers.prototype.upsertUser = function (user, options, callback) {
 
   try {
     _this.__validate('user', options, user, function (e) {
-      if (e) return callback(e);
 
+      if (e) return callback(e);
       _this.__upsertUser(user, options, callback);
     });
   } catch (e) {
@@ -343,195 +228,18 @@ SecurityUsers.prototype.deleteUser = function (user, options, callback) {
   });
 };
 
-SecurityUsers.prototype.__decodeGroupPermissions = function (group) {
+SecurityUsers.prototype.__removeGroupLinks = function (userItems) {
+  var usersList = [];
 
-  var parsedPermissions = {};
+  for (var itemIndex in userItems) {
 
-  for (var permissionKey in group.permissions) {
-    parsedPermissions[permissionKey.replace(/\\u002e/, ".")] = group.permissions[permissionKey];
+    var userItem = userItems[itemIndex];
+    if (userItem._meta['path'].indexOf('_USER_GROUP') >= 0) continue;
+    usersList.push(userItem);
+
   }
 
-  group.permissions = parsedPermissions;
-
-  return group;
-
-};
-
-SecurityUsers.prototype.__encodeGroupPermissions = function (group) {
-
-  var parsedPermissions = {};
-
-  for (var permissionKey in group.permissions) {
-    parsedPermissions[permissionKey.replace(/\./g, "\\u002e")] = group.permissions[permissionKey];
-  }
-
-  group.permissions = parsedPermissions;
-
-  return group;
-
-};
-
-SecurityUsers.prototype.__upsertGroup = function (group, options, callback) {
-
-  var groupPath;
-
-  if (options.parent) groupPath = options.parent._meta.path + '/' + group.name;
-  else groupPath = '/_SYSTEM/_SECURITY/_GROUP/' + group.name;
-
-  var _this = this;
-
-  _this.dataService.upsert(groupPath, group, {
-    merge: true
-  }, function (e, result) {
-
-    if (e) return callback(e);
-
-    var upserted = _this.serialize('group', result);
-
-    _this.securityService.dataChanged('upsert-group', upserted, null, function(){
-
-      callback(null, upserted);
-    });
-
-  }.bind(this));
-
-};
-
-SecurityUsers.prototype.upsertGroup = function (group, options, callback) {
-
-  if (typeof options == 'function') callback = options;
-
-  var preparedGroup = this.__encodeGroupPermissions(group);
-  //TODO: this was done, but the variable never used - now testing using it
-  try {
-    this.__validate('group', options, preparedGroup, function (e) {
-      if (e) return callback(e);
-      this.__upsertGroup(preparedGroup, options, callback);
-    }.bind(this));
-  } catch (e) {
-    callback(e);
-  }
-};
-
-SecurityUsers.prototype.deleteGroup = function (group, options, callback) {
-
-  var _this = this;
-
-  if (typeof options == 'function') callback = options;
-
-  _this.getGroup(group.name, {}, function (e, group) {
-
-    if (e) return callback(e);
-
-    if (!group) return callback(new Error('group you are deleting does not exist'));
-
-    var deletePath = '/_SYSTEM/_SECURITY/_USER/*/_USER_GROUP/' + group.name + '/*';
-
-    try {
-
-      _this.dataService.remove(deletePath, {}, function (e, userGroupDeleteResults) {
-
-        if (e) return callback(e);
-
-        deletePath = '/_SYSTEM/_SECURITY/_GROUP/' + group.name;
-
-        _this.dataService.remove(deletePath, {}, function (e, groupDeleteResults) {
-
-          if (e) return callback(e);
-
-          var deleted = {
-            removed: groupDeleteResults.data.removed,
-            obj: group,
-            links: userGroupDeleteResults
-          };
-
-          _this.securityService.dataChanged('delete-group', deleted, null, function(){
-
-            callback(null, deleted);
-          });
-        });
-      });
-    } catch (e) {
-      callback(e);
-    }
-  });
-};
-
-SecurityUsers.prototype.listGroups = function (groupName, options, callback) {
-
-  if (typeof options == 'function') callback = options;
-
-  if (typeof groupName == 'function') {
-    callback = groupName;
-    groupName = '*';
-  }
-
-  if (groupName[groupName.length - 1] != '*') groupName += '*';
-
-  var searchPath = '/_SYSTEM/_SECURITY/_GROUP/' + groupName;
-
-  try {
-
-    this.dataService.get(searchPath, {
-      sort: {
-        'path': 1
-      }
-    }, function (e, results) {
-
-      if (e) return callback(e);
-
-      callback(null, this.serializeAll('group', results, options));
-
-    }.bind(this));
-  } catch (e) {
-
-    callback(e);
-  }
-};
-
-SecurityUsers.prototype.getGroup = function (groupName, options, callback) {
-
-  if (typeof options == 'function') callback = options;
-
-  groupName = groupName.replace(/[*]/g, '');
-
-  var searchPath = '/_SYSTEM/_SECURITY/_GROUP/' + groupName;
-
-  var _this = this;
-
-  try {
-
-    _this.__cache_groups.get(groupName, function (e, cachedGroup) {
-
-      if (e) return callback(e);
-
-      if (cachedGroup) return callback(null, cachedGroup);
-
-      _this.dataService.get(searchPath, {
-        sort: {
-          'path': 1
-        }
-      }, function (e, result) {
-
-        if (e) return callback(e);
-
-        if (result) {
-
-          var group = _this.__decodeGroupPermissions(_this.serialize('group', result, options));
-
-          //add the group to the cache TODO: implement LRU cache service here
-          _this.__cache_groups.set(groupName, group)
-            .then(callback(null, group))
-            .catch(callback);
-
-        } else callback(null, null);
-
-      });
-    });
-
-  } catch (e) {
-    callback(e);
-  }
+  return usersList;
 };
 
 SecurityUsers.prototype.getUser = function (userName, options, callback) {
@@ -576,7 +284,7 @@ SecurityUsers.prototype.getUser = function (userName, options, callback) {
 
           });
 
-          var returnUser = _this.serialize('user', user, options);
+          var returnUser = _this.securityService.serialize('user', user, options);
 
           if (groups) returnUser.groups = groups;
 
@@ -592,20 +300,6 @@ SecurityUsers.prototype.getUser = function (userName, options, callback) {
   } catch (e) {
     callback(e);
   }
-};
-
-SecurityUsers.prototype.__removeGroupLinks = function (userItems) {
-  var usersList = [];
-
-  for (var itemIndex in userItems) {
-
-    var userItem = userItems[itemIndex];
-    if (userItem._meta['path'].indexOf('_USER_GROUP') >= 0) continue;
-    usersList.push(userItem);
-
-  }
-
-  return usersList;
 };
 
 SecurityUsers.prototype.listUsers = function (userName, options, callback) {
@@ -624,84 +318,17 @@ SecurityUsers.prototype.listUsers = function (userName, options, callback) {
 
   try {
 
-    this.dataService.get(searchPath, {}, function (e, results) {
+    var _this = this;
+
+    _this.dataService.get(searchPath, {}, function (e, results) {
 
       if (e) return callback(e);
 
-      var users = this.__removeGroupLinks(results); //had to do this to be compatible with mongo
+      var users = _this.__removeGroupLinks(results); //had to do this to be compatible with mongo
 
-      callback(null, this.serializeAll('user', users, options));
-
-    }.bind(this));
-
-  } catch (e) {
-    callback(e);
-  }
-};
-
-SecurityUsers.prototype.linkGroup = function (group, user, options, callback) {
-
-  var _this = this;
-
-  if (typeof options == 'function') {
-    callback = options;
-    options = {};
-  }
-
-  try {
-
-    _this.__validate('user-group', [group, user], options, function (e) {
-
-      if (e) return callback(e);
-
-      var groupLinkPath = '/_SYSTEM/_SECURITY/_USER/' + user.username + '/_USER_GROUP/' + group.name;
-
-      _this.dataService.upsert(groupLinkPath, options, {
-        merge: true
-      }, function (e, result) {
-
-        if (e) return callback(e);
-
-        var upserted = _this.serialize('user-group', result, options);
-
-        _this.securityService.dataChanged('link-group', upserted, user, function(){
-
-          callback(null, upserted);
-        });
-      });
+      callback(null, _this.securityService.serializeAll('user', users, options));
     });
-  } catch (e) {
-    callback(e);
-  }
-};
 
-SecurityUsers.prototype.unlinkGroup = function (group, user, options, callback) {
-
-  var _this = this;
-
-  if (typeof options == 'function') {
-    callback = options;
-    options = {};
-  }
-
-  try {
-
-    _this.__validate('user-group', [group, user], null, function (e) {
-
-      if (e) return callback(e);
-
-      var groupLinkPath = '/_SYSTEM/_SECURITY/_USER/' + user.username + '/_USER_GROUP/' + group.name;
-
-      _this.dataService.remove(groupLinkPath, {}, function (e, result) {
-
-        if (e) return callback(e);
-
-        _this.securityService.dataChanged('unlink-group', groupLinkPath, user, function(){
-
-          callback(null, result);
-        });
-      });
-    });
   } catch (e) {
     callback(e);
   }
